Updated the Annotate Coloumn
Made the annotate column into two seperate components, one for makring
up the page, and one for annotating the page... allowing for the analyze
page to be flexible
