+++ citing-insights/src/Markup.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react'
+import { Label, ListGroup, ListGroupItem, Button, Input, Jumbotron, Badge, Progress } from 'reactstrap';
+
+
+class Markup extends Component {
+
+
+  render() {
+    return(
+    	<div class="markup-container">
+    		<h2>Load Markup Component here: </h2>
+    		<p>This component allows faculty and staff to select and mark intext citations</p>
+	    	
+		</div>
+    );
+  }
+}
+
+export default Markup;
